Move all logic from app to main
We can use context API in main now
