+++ hooks/index.js
@@ -1,2 +1,2 @@
 export { default as useAnalytics, usePageviewTracking } from './analytics'
-export { default as useCookies } from './cookies'
+export default from './cookies'
